+++ test/mocha/mock.data.js
@@ -1,5 +1,5 @@
 /*!
- * Copyright (c) 2017-2019 Digital Bazaar, Inc. All rights reserved.
+ * Copyright (c) 2017-2021 Digital Bazaar, Inc. All rights reserved.
  */
 'use strict';
 
@@ -12,12 +12,16 @@ const mock = {};
 module.exports = mock;
 
 mock.ledgerConfiguration = {
-  '@context': constants.WEB_LEDGER_CONTEXT_V1_URL,
+  '@context': [
+    constants.WEB_LEDGER_CONTEXT_V1_URL,
+    constants.ED25519_2020_CONTEXT_V1_URL
+  ],
   type: 'WebLedgerConfiguration',
   ledger: 'did:v1:eb8c22dc-bde6-4315-92e2-59bd3f3c7d59',
   consensusMethod: 'Continuity2017',
-  electorSelectionMethod: {
-    type: 'MostRecentParticipants',
+  witnessSelectionMethod: {
+    type: 'WitnessPoolWitnessSelection',
+    witnessPool: 'did:v1:uuid:2f3c9466-ddc9-11eb-92f2-f31707920b3b'
   },
   sequence: 0
 };